Empty scene, a proto level and player adjustments
I made a Empty scene, a derived prototype level.
The skeleton for a teleporter pads(need to integrate with trigger event)
Adjusted the player script so that the joint costraints can be tuned
while in the editor. Also the procedure can be used in game, so to
switch different behavior.